+++ cosmovisor/scanner_test.go
@@ -14,39 +14,53 @@ func TestParseUpgradeInfoFile(t *testing.T) {
 		filename      string
 		expectUpgrade upgradetypes.Plan
 		expectErr     bool
-	}{{
-		filename:      "f1-good.json",
-		expectUpgrade: upgradetypes.Plan{Name: "upgrade1", Info: "some info", Height: 123},
-		expectErr:     false,
-	}, {
-		filename:      "f2-bad-type.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}, {
-		filename:      "f2-bad-type-2.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}, {
-		filename:      "f3-empty.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}, {
-		filename:      "f4-empty-obj.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}, {
-		filename:      "f5-partial-obj-1.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}, {
-		filename:      "f5-partial-obj-2.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}, {
-		filename:      "unknown.json",
-		expectUpgrade: upgradetypes.Plan{},
-		expectErr:     true,
-	}}
+	}{
+		{
+			filename:      "f1-good.json",
+			expectUpgrade: upgradetypes.Plan{Name: "upgrade1", Info: "some info", Height: 123},
+			expectErr:     false,
+		},
+		{
+			filename:      "f2-normalized-name.json",
+			expectUpgrade: upgradetypes.Plan{Name: "upgrade2", Info: "some info", Height: 125},
+			expectErr:     false,
+		},
+		{
+			filename:      "f2-bad-type.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+		{
+			filename:      "f2-bad-type-2.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+		{
+			filename:      "f3-empty.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+		{
+			filename:      "f4-empty-obj.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+		{
+			filename:      "f5-partial-obj-1.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+		{
+			filename:      "f5-partial-obj-2.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+		{
+			filename:      "unknown.json",
+			expectUpgrade: upgradetypes.Plan{},
+			expectErr:     true,
+		},
+	}
 
 	for i := range cases {
 		tc := cases[i]