+++ tx/textual/valuerenderer/bytes_test.go
@@ -2,9 +2,9 @@ package valuerenderer_test
 
 import (
 	"context"
-	"encoding/hex"
+	"encoding/base64"
 	"encoding/json"
-	"io/ioutil"
+	"os"
 	"strings"
 	"testing"
 
@@ -12,34 +12,42 @@ import (
 	"google.golang.org/protobuf/reflect/protoreflect"
 )
 
-func TestFormatBytes(t *testing.T) {
+func TestBytesJsonTestCases(t *testing.T) {
 	var testcases []bytesTest
-	raw, err := ioutil.ReadFile("../internal/testdata/bytes.json")
+	// Bytes.json contains bytes that are represented in base64 format, and
+	// their expected results in hex.
+	raw, err := os.ReadFile("../internal/testdata/bytes.json")
 	require.NoError(t, err)
 
 	err = json.Unmarshal(raw, &testcases)
 	require.NoError(t, err)
 
 	for _, tc := range testcases {
-		data, err := hex.DecodeString(tc.hex)
+		data, err := base64.StdEncoding.DecodeString(tc.base64)
 		require.NoError(t, err)
 
-		r, err := valueRendererOf(data)
+		valrend, err := valueRendererOf(data)
 		require.NoError(t, err)
 
 		b := new(strings.Builder)
-		err = r.Format(context.Background(), protoreflect.ValueOfBytes(data), b)
+		err = valrend.Format(context.Background(), protoreflect.ValueOfBytes(data), b)
 		require.NoError(t, err)
-		require.Equal(t, tc.expRes, b.String())
+		require.Equal(t, tc.hex, b.String())
+
+		// Round trip
+		r := strings.NewReader(tc.hex)
+		val, err := valrend.Parse(context.Background(), r)
+		require.NoError(t, err)
+		require.Equal(t, tc.base64, base64.StdEncoding.EncodeToString(val.Bytes()))
 	}
 }
 
 type bytesTest struct {
 	hex    string
-	expRes string
+	base64 string
 }
 
 func (t *bytesTest) UnmarshalJSON(b []byte) error {
-	a := []interface{}{&t.hex, &t.expRes}
+	a := []interface{}{&t.hex, &t.base64}
 	return json.Unmarshal(b, &a)
 }